All Pages component works with react-redux (w/o actual database)
