+++ client/routes.js
@@ -3,7 +3,7 @@ import {connect} from 'react-redux'
 import {withRouter, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
 import {Login, Signup, UserHome} from './components'
-import {AllProducts} from './components/All-Products'
+import {default as AllProducts} from './components/All-Products'
 import {me} from './store'
 import {HomePage} from './components/homepage'
